% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{ci.cvAUC_withIC}
\alias{ci.cvAUC_withIC}
\title{ci.cvAUC_withIC}
\usage{
ci.cvAUC_withIC(
  predictions,
  labels,
  label.ordering = NULL,
  folds = NULL,
  confidence = 0.95
)
}
\description{
This function is taken nearly verbatim from the \code{cvAUC} package. The only difference
is that it additionally returns estimated influence functions, which allows
us to do later comparisons of CV-AUC between different learners.
}
