% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{diff_cvAUC}
\alias{diff_cvAUC}
\title{diff_cvAUC}
\usage{
diff_cvAUC(fit1, fit2, confidence = 0.95)
}
\arguments{
\item{fit1}{An object of class \code{wrap_AUC}}

\item{fit2}{An object of class \code{wrap_AUC}}

\item{confidence}{A \code{numeric} between 0 and 1 indicating the nominal coverage
probability for the confidence interval.}
}
\value{
An object of class \code{diff_cvAUC} with the following entries:
\item{diff}{The difference in CV-AUC between the two fits.}
\item{ci}{The confidence interval for the difference between the two fits. }
\item{p}{The two-sided p-value for the test that the two CV-AUCs are equal. }
\item{folds}{The number of folds used by \cote{fit1} and \code{fit2}.}
\item{learner1}{The name of the learner used for \code{fit1}.}
\item{learner2}{The name of the learner used for \code{fit2}.}
\item{confidence}{The confidence interval level.}
}
\description{
This function estimates the difference between two cross-validated AUCs
fit using the \code{wrap_cvAUC} function. It also provides an influence 
function-based confidence interval estimate and hypothesis test of the null
hypothesis that the two CV-AUCs are equal.
}

