% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{.getPredictions}
\alias{.getPredictions}
\title{.getPredictions}
\usage{
.getPredictions(learner, Y, X, V, folds, returnFits, parallel)
}
\description{
This function calls the specified \code{learner} over each fold and
returns a list of cross-validated predictions and model fits.
}
