% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cvROC.R
\name{plot_cvROC}
\alias{plot_cvROC}
\title{plot_cvROC}
\usage{
plot_cvROC(Y, X, wrap_cvAUC_fit, plotAverage = TRUE, ...)
}
\arguments{
\item{Y}{A \code{numeric} vector of class labels}

\item{X}{A \code{data.frame} of variables that was used in the call to \code{wrap_cvAUC}}

\item{wrap_cvAUC_fit}{An object of class \code{wrap_cvAUC}. It is assumed that \code{Y}
is in the same order as when \code{wrap_cvAUC} was called.}

\item{plotAverage}{A \code{boolean} indicating whether to plot the average ROC curve
over cross-validation folds.}

\item{...}{Other options passed to \code{plot}.}
}
\description{
Plot cross-validated ROC curves and the averaged ROC curve.
}
