% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{print.wrap_cvAUC}
\alias{print.wrap_cvAUC}
\title{print.wrap_cvAUC}
\usage{
\method{print}{wrap_cvAUC}(object, digits = 2)
}
\arguments{
\item{object}{An object of class \code{wrap_cvAUC}.}

\item{digits}{A \code{numeric} providing the number of digits to round to.}
}
\description{
A print method for objects of class \code{wrap_cvAUC}
}

